requirements.txt
flask==3.0.0
pillow==10.1.0
qrcode==7.4.2
tk==0.1.0